ATO-647 - make consistent type for the NTrCut+warning removal + changing index

+++ Detectors/AOD/include/AODProducerWorkflow/AODProducerWorkflowSpec.h
@@ -230,7 +230,7 @@ class AODProducerWorkflowDPL : public Task
   bool mPropTracks{false};
   bool mPropMuons{false};
   float mTrackQCFraction{0.05};
-  uint8_t mTrackQCNTrCut{4};
+  int64_t mTrackQCNTrCut{4};
   float mSqrtS{13860.};
   std::mt19937 mGenerator; ///< random generator for trackQA sampling
   o2::base::Propagator::MatCorrType mMatCorr{o2::base::Propagator::MatCorrType::USEMatCorrLUT};
